// src/api/type/ProductType.ts
export type ProductType = {
  product: object;
  quantity: number;
  id: number;
  category: string;
  itemId: string;
  name: string;
  fullPrice: number;
  price: number;
  screen: string;
  capacity: string;
  color: string;
  ram: string;
  year: number;
  image: string;
  images: string;
  priceDiscount: number;
  priceRegular: number;
};